+++ src/main/java/com/jk18_7/sim/authority/service/RolePermissionService.java
@@ -3,7 +3,7 @@ package com.jk18_7.sim.authority.service;
 import com.jk18_7.sim.authority.entity.Role;
 
 public interface RolePermissionService {
-    void setPermissions(Role role, long[] pids);
-    void deleteByRole(long rid);
-    void deleteByPermission(long pid);
+    void setPermissions(Role role, int[] pids);
+    void deleteByRole(int rid);
+    void deleteByPermission(int pid);
 }
