+++ src/main/java/com/jk18_7/sim/authority/service/RoleService.java
@@ -12,7 +12,7 @@ public interface RoleService {
     List<Role> list(Users users);
     List<Role> list();
     void addRole(Role role);
-    void deleteRole(long id);
-    Role getRole(long id);
+    void deleteRole(int id);
+    Role getRole(int id);
     void updateRole(Role role);
 }
